server/requirements.txt
openai
python-dotenv
langchain-text-splitters
chromadb
gitpython